branch: HEAD speed up mdsrate test. b=20501 r=rread r=grev

+++ lustre/tests/test-framework.sh
@@ -2891,7 +2891,10 @@ get_mds_dir () {
 }
 
 mdsrate_cleanup () {
-    mpi_run -np $1 -machinefile $2 ${MDSRATE} --unlink --nfiles $3 --dir $4 --filefmt $5 $6
+    if [ -d $4 ]; then
+        mpi_run -np $1 -machinefile $2 ${MDSRATE} --unlink --nfiles $3 --dir $4 --filefmt $5 $6
+        rmdir $4
+    fi
 }
 
 delayed_recovery_enabled () {
